+++ CHANGES.rst
@@ -3,8 +3,8 @@ Change Log
 
 Async PRAW follows `semantic versioning <http://semver.org/>`_.
 
-Unreleased
-----------
+7.4.0 (2021/07/30)
+------------------
 
 **Added**